Fix clang-tidy warnings
Resolve every finding the lint workflow reports, across both the native
and terminal translation units:

- actor.c: cast current_position to int before centering so the sprite
  offset isn't an integer division in a float context
  (bugprone-integer-division).
- intro.c: drop two unused, underscore-prefixed chunk pointers
  (bugprone-reserved-identifier + unused-variable).
- playground.c: use expf() instead of exp() on a float
  (performance-type-promotion-in-math-fn).
- game.h: name the register_adventures() parameter to match its
  definition (readability-inconsistent-declaration-parameter-name).
- image.c: split the comma-separated declarations in the .anim parser
  (readability-isolate-declaration).
- main_terminal.c: brace the ESC handler and drop the lowercase 'f'
  literal suffix (readability-braces-around-statements,
  readability-uppercase-literal-suffix).
- terminal.c: brace four single-line guards and widen the pixel-buffer
  size to size_t before multiplying (readability-braces-around-statements,
  bugprone-implicit-widening-of-multiplication-result).

clang-tidy now reports zero warnings (the CI count was higher because
`bear -- make all terminal` records every shared source twice, once per
build, so each warning was double-counted; fixing the source clears
both). All three targets build clean and the headless web smoke test
plays the full Gina puzzle (hub -> dive) with no page errors.

+++ src/actor.c
@@ -158,8 +158,8 @@ void actor_render(Actor *actor, SDL_Renderer *renderer) {
 
   // Assumes that all animation frames have the same size.
   SDL_Point position = (SDL_Point){
-      .x = actor->current_position.x - reference->sprite_clips[0].w / 2,
-      .y = actor->current_position.y - reference->sprite_clips[0].h / 2};
+      .x = (int)actor->current_position.x - reference->sprite_clips[0].w / 2,
+      .y = (int)actor->current_position.y - reference->sprite_clips[0].h / 2};
 
   // IDLE renders the actor's idle animation (e.g. the fox's sitting sprite).
   ActorState state =
